api/requirements.txt
djangorestframework == 3.13.1
drf-yasg == 1.20.0